Restructurization
files cleanup
