+++ src/main/java/com/emotion/emotiondesktop/Controller/CitiesCrudController.java
@@ -1,7 +1,13 @@
-package com.emotion.emotiondesktop;
+package com.emotion.emotiondesktop.Controller;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
+
+import com.emotion.emotiondesktop.EmotionApplication;
+import com.emotion.emotiondesktop.Helper.HttpRequest;
+import com.emotion.emotiondesktop.Helper.HttpResponse;
+import com.emotion.emotiondesktop.Helper.Validation;
+import com.emotion.emotiondesktop.Model.City;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
