+++ src/main/java/com/emotion/emotiondesktop/Controller/MapController.java
@@ -1,6 +1,8 @@
-package com.emotion.emotiondesktop;
+package com.emotion.emotiondesktop.Controller;
 
 import java.io.IOException;
+
+import com.emotion.emotiondesktop.EmotionApplication;
 import com.sothawo.mapjfx.*;
 import javafx.fxml.FXML;
 public class MapController {
