+++ src/main/java/com/emotion/emotiondesktop/Helper/HttpResponse.java
@@ -1,4 +1,4 @@
-package com.emotion.emotiondesktop;
+package com.emotion.emotiondesktop.Helper;
 
 import lombok.Getter;
