+++ src/main/resources/com/emotion/emotiondesktop/cities-crud-view.fxml
@@ -18,7 +18,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane style="-fx-background-color: ffffff;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.emotion.emotiondesktop.CitiesCrudController">
+<AnchorPane style="-fx-background-color: ffffff;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.emotion.emotiondesktop.Controller.CitiesCrudController">
    <children>
       <BorderPane prefHeight="768.0" prefWidth="1366.0">
          <left>
