+++ src/main/resources/com/emotion/emotiondesktop/countries-crud-view.fxml
@@ -18,7 +18,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane style="-fx-background-color: ffffff;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.emotion.emotiondesktop.CountriesCrudController">
+<AnchorPane style="-fx-background-color: ffffff;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.emotion.emotiondesktop.Controller.CountriesCrudController">
     <children>
         <BorderPane prefHeight="768.0" prefWidth="1366.0">
             <left>
