+++ src/main/resources/com/emotion/emotiondesktop/login-view.fxml
@@ -12,7 +12,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane prefHeight="768.0" prefWidth="1366.0" style="-fx-background-color: #f4f8fd;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.emotion.emotiondesktop.LoginController">
+<AnchorPane prefHeight="768.0" prefWidth="1366.0" style="-fx-background-color: #f4f8fd;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.emotion.emotiondesktop.Controller.LoginController">
    <children>
       <VBox alignment="CENTER" fillWidth="false" layoutX="361.0" layoutY="182.0" prefHeight="450.0" prefWidth="600.0" spacing="120.0" style="-fx-background-color: #ffffff; -fx-opacity: 80%;">
          <children>
