+++ src/test/java/com/emotion/emotiondesktop/Integration/LoginControllerTest.java
@@ -1,11 +1,10 @@
 package com.emotion.emotiondesktop.Integration;
 
-import com.emotion.emotiondesktop.LoginController;
+import com.emotion.emotiondesktop.Controller.LoginController;
 import javafx.application.Platform;
 import javafx.scene.control.Label;
 import javafx.scene.control.PasswordField;
 import javafx.scene.control.TextField;
-import javafx.stage.Stage;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
