+++ src/test/java/com/emotion/emotiondesktop/Unit/HttpRequestTest.java
@@ -1,0 +1,6 @@
+package com.emotion.emotiondesktop.Unit;
+
+class HttpRequestTest {
+
+
+}
